Agent work for MUNIDESK-1421: ADR 0012 macOS activation: register the bundled LaunchAgent

+++ src-tauri/src/macos_runtime_service.rs
@@ -126,7 +126,7 @@ impl RuntimeServiceAdapter for MacosRuntimeServiceAdapter {
 
         // SAFETY: The service comes from agentServiceWithPlistName and remains retained.
         unsafe { self.service.registerAndReturnError() }.map_err(|error| {
-            if error.code() == i64::from(kSMErrorLaunchDeniedByUser) {
+            if error.code() == kSMErrorLaunchDeniedByUser {
                 RegistrationError::Denied
             } else {
                 RegistrationError::Failed
